requirements.txt
alembic==1.13.1
bcrypt==4.1.3
blinker==1.8.2
Bootstrap-Flask==2.4.0
click==8.1.7
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
python-dotenv==1.0.1
SQLAlchemy==2.0.25
typing_extensions==4.12.0
Werkzeug==3.0.0
WTForms==3.1.2
